A complete rewrite of the threeplayer.py plugin
A total rewrite of the threeplayer.py file to bring it in line with the
format of the other onDemand plugins. New method to parse the html using
lxml so a new dependancy for this to be installed also.

The plugin still needs work for the findPlayUrl function, this is very
slow but I could not figure out a quicker way to get this to run.