[onDemand] Fix threeplayer.py skipping shows
When pulling the media data I was looking for div's that had and id =
"gridshow" instead of an id containing "gridshow". This resulted in
every 5th show getting dropped (Good spot Sean)

Also, putting in some of Seans code to remove the call to BeautifulSoup
in the function findPlayUrl so this should help speed up getting the
stream going.