[OnDemand] threeplayer.py fix Straight Stream
The html that was being returned for the Striahgt off the Telly streams
had changed so I needed to change the findPlayUrl function to retrieve
the mp4 url.

Changed it slightly also so that it should future proof it a bit if it
changes again in the future.

Also added a few MessageBox popup's to display an error if this happens
in future.