[OnDemand] BBC iPlayer Play Stream Update
Currently if we don't find the preferred stream quality or lower an
error message is thrown and nothing is played even if a playable stream
is found.

We are now going to play the available stream now anyway even if it
doesn't match the preferred quality setting but a message will be
displayed to notify you that this has happened.