[LCD4linux] improvement v5.0-r9
- improvement for displaying covers
Now user can add a list of filenames for covers. This opens the possibility for several 3rd-party plugins to deposit coverfiles in a folder (e.g. /tmp/mycover.png) and enables LCD4linux to find and display it.

example usage: 'mycover.jpg, mycover2.jpg, myfilename.*'

In case the extension '*' (= joker) is given, LCD4linux will search for 'myfilename.jpg' and when it fails, LCD4linux will search for 'myfilename.png'